+++ test/rails_app/app/models/account.rb
@@ -1,4 +1,9 @@
 class Account < ApplicationRecord
-  validates_presence_of :email
   include Rodauth::Rails.model
+  if ActiveRecord.version >= Gem::Version.new("7.0")
+    enum :status, unverified: 1, verified: 2, closed: 3
+  else
+    enum status: { unverified: 1, verified: 2, closed: 3 }
+  end
+  validates_presence_of :email
 end